sally/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
together>=1.5.0
httpx==0.27.0
pydantic>=2.6.3
python-multipart==0.0.6
Pillow>=11.1.0
requests==2.31.0
python-dotenv==1.0.0 